added main.cpp, a lot
did a whole bunch of code tidy up.

menu is now located in a menu function that calls the choice function. this is called in main().

returnMenu is a function that asks the user if they would like to return to the main menu (y/n). y works, n goes to readWord() function because thats what it is told to do. If anyone has any ideas on how to make it go back to the previously selected function, please input.

Just copy and paste the code from the cpp file to yours if need.

this code is running.